Fixed dependency on printer control variables for internally generated names.
The internally created layered function names (and other names) depended on printer control variables, such as *print-case*. This can lead to problems when user programs set them globally. This is fixed by wrapping the generation of internal names in a with-standard-io-syntax form.

Thanks a lot to Kambiz Darabi for reporting the issue.